reduce copy pasting in rhat
